chore: move env constants
