+++ src/components/Sidebar/index.tsx
@@ -16,13 +16,14 @@ import {
 } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import { ExpandMore } from '@material-ui/icons';
-import { SITE_VERSION, NETWORK_LIST, FLAGS } from 'config/constants';
+import { SITE_VERSION, NETWORK_LIST } from 'config/constants';
 import NetworkWidget from 'components-v2/common/NetworkWidget';
 import { Route } from 'mobx-router';
 import { RootStore } from 'mobx/RootStore';
 import clsx, { ClassValue } from 'clsx';
 import SecurityIcon from '@material-ui/icons/Security';
 import { sidebarPricingLinks } from 'config/ui/links';
+import { FLAGS } from 'config/environment';
 
 const DRAWER_WIDTH = 240;
