+++ src/config/networks.config.ts
@@ -3,7 +3,7 @@ import { Ethereum } from 'mobx/model/network/eth.network';
 import { Polygon } from 'mobx/model/network/matic.network';
 import { Network } from 'mobx/model/network/network';
 import { xDai } from 'mobx/model/network/xdai.network';
-import { FLAGS } from './constants';
+import { FLAGS } from './environment';
 
 export const supportedNetworks: Network[] = [
 	new Ethereum(),
