+++ src/mobx/utils/helpers.ts
@@ -1,5 +1,5 @@
 import BigNumber from 'bignumber.js';
-import { DEBUG, getDefaultRetryOptions, TEN, ZERO } from '../../config/constants';
+import { getDefaultRetryOptions, TEN, ZERO } from '../../config/constants';
 import { API } from 'bnc-onboard/dist/src/interfaces';
 import store from 'mobx/RootStore';
 import { retry } from '@lifeomic/attempt';
@@ -7,6 +7,7 @@ import { MarketChartStats } from 'mobx/model/charts/market-chart-stats';
 import { MarketDelta } from 'mobx/model/charts/market-delta';
 import { ChartData } from 'mobx/model/charts/chart-data';
 import { Network } from 'mobx/model/network/network';
+import { DEBUG } from 'config/environment';
 
 export const jsonQuery = (url: string | undefined): Promise<Response> | undefined => {
 	if (!url) return;
